Added #[allow(unstable)] because of libc, test-suite and slices
We need to use libc for the external bindings, and libc isn't stable as
of yet.

Also for now slicing syntax doesn't seem to be working for most of the code
that takes slices.

The test-suite uses the test-crate which also isn't stable at the
moment, it also uses the serialize-crate for some of the test vectors.
We are supposed to use rustc-serialize instead, but for now I'm
reluctant to add a test-only dependency to sodiumoxide, maybe if I
figure out how to only make the test-suite depend on the rustc-serialize
crate.