Replaced `.as_slice()` with Index<Range> et.al.
Since `.as_slice()` is deprecated for rusts own types we implement
Index<Range>, Index<RangeTo>, Index<RangeFrom> and Index<RangeFull> for
all types so that slice syntax is used for all types in sodiumoxide.

For now we keep `.from_slice()` and we don't implement the
IndexMut-traits. Will there be a need to implement those?