Simplify uses of macros
Now that 'use' statements don't need to be at the beginning of files
anymore we can have use statements in our macro definitions instead of
in the files that call the macros. This simplifies macro invocation a
lot and cleans up the imports in a lot of files.

However, this makes the macros a bit unhygienic, since a use-statement
can't occur twice in calling code. On the other hand they were already
unhygienic since they needed things to be in scope to work.