Remove derived Copy for private key
Appears we cannot derive a Copy trait when we define a drop method.

We make Scalar work exactly like every other private key value in
sodiumoxide with this change.

I'm bothered by public and private keys not being both copyable or
not together, not sure if there is a good solution though.