Add further tests for crypto_generichash_blake2b
Add tests for the bindings of:
* crypto_generichash_blake2b
* crypto_generichash_blake2b_salt_personal

These tests only check that the bindings can be called and report success by
returning zero where appropriate. The hash algorithm itself is not tested here.

The following functions cannot yet be tested:
* crypto_generichash_blake2b_init
* crypto_generichash_blake2b_init_salt_personal
* crypto_generichash_blake2b_update
* crypto_generichash_blake2b_final

To test these functions we need to use crypto_generichash_blake2b_statebytes
which has only just been added to libsodium. We should wait until it is included
in a release before using it here.