implement detached and in-place box and secretbox functions
https://github.com/dnaq/sodiumoxide/issues/24

This is another take on non-allocating primitives for box and secretbox.
There were two particular design questions in my head for what interface
to expose for this:

1. Combined mode vs detached mode. Detached mode is the lower-level
primitive of the two, and it's easy to implement combined mode as a
wrapper, but not the other way around. This is also how libsodium's
implementation is organized. So I went with detached mode.

2. In-place vs two-buffers. Libsodium supports both, by checking the
pointers to see if they're overlapping. Rust makes it hard to support
both in a single interface, because &mut references can never be
aliased. I think it's easier to wrap an in-place function to become a
two-buffers one than the other way around, so I went with in-place.

Detached-and-in-place has the extra advantage that it can never panic,
because there are no length checks. The length of the tag is statically
known, and any length of the message is valid.

https://github.com/dnaq/sodiumoxide/pull/150