Remove `crypto::stream::salsa208` and `crypto::stream::salsa2012`
This is in preparation for #181. We want the next commit to add
`ic` versions of all stream ciphers, but in libsodium neither
`crypto_stream_salsa208` nor `crypto_stream_salsa2012` exposes functions
were we are allowed to set the initial block counter.

Since we wan't the API to be identical for all stream ciphers the
easiest thing to do is to remove these ciphers. No-one should be using
these constructions anyway.