Format all code using cargo fmt
This is a change that has been long overdue and it should simplify for
others wanting to contribute to the repo.

Unfortunately it touches a lot of files, which in turn might lead to
some unnecessary merge conflicts for existing pull-requests. On the
other hand it's not really that hard for those contributors to run cargo
fmt on their code as well.