Avoid defining unused nonce generation functions
This adds a generic function to create random nonces which is not accessible
outside of this crate and only exposed where this is safe due to
the length of the nonces being sufficiently large.

To do so, a `Nonce` trait is created that is also not visible to dependent crates
but allows generic construction from an array of bytes without extending the
public interface of the nonce types.