Use `ed25519::Signature` type; impl `Signer`/`Verifier` traits
The `ed25519` crate provides a common `Signature` type which can be
shared among libraries which implement Ed25519 signing/verification.

Without common types/traits for Ed25519 signatures, downstream consumer
crates which use Ed25519 as part of a high-level protocol implementation
which allows consumers of their crates to be able to use their choice of
Rust Ed25519 libraries wind up writing their own abstractions to do so
on a crate-by-crate basis.

The `ed25519` crate, and the `signature` crate it's built on, provide
a common abstraction layer:

- `ed25519` crate: https://github.com/RustCrypto/signatures/tree/master/ed25519
- `signature` crate: https://github.com/RustCrypto/traits/tree/master/signature

The `ed25519` crate has been natively adopted by `ed25519-dalek`:

- Issue: https://github.com/dalek-cryptography/ed25519-dalek/issues/80
- Usage example: https://docs.rs/ed25519/1.0.2/ed25519/#using-above-example-with-ed25519-dalek

If `sodiumoxide` were to adopt these types/traits as well, I'd be happy
to add similar documentation to the `ed25519` crate for how to use them
with `sodiumoxide`.