secretstream: More general push/pull implementations
This commit changes the implementation of `crypto::secretstream::Stream`
to use the methods `push_impl_ptr` and `pull_impl_ptr` for push/pull,
which support using an arbitrary pointer as the destination for the
encrypted/decrypted result, rather than just a vector. The intention of
this commit is to lay the groundwork for implementing
`Stream.push_to_slice` and `Stream.pull_to_slice`.