Update pre-built libraries to latest stable
This commit updates the pre-built Windows libraries (MinGW and MSVC) to
the latest stable 1.0.18 release from https://download.libsodium.org/.
In addition, it includes the compressed archives from which the
libraries are extracted, and the signatures for these archives, signed
by jedisct1 (Frank Denis, the author of libsodium) using Minisign. This
allows users to verify the libraries we distribute in sodiumoxide are
those distributed from the official website.